bucky list: retry-able HTTP requests
If the HTTP request has a body (like a POST) the first attempt to issue
that request will drain the body and further attempts will result in an
error as the content-length is set to longer than the 0 byte body.

We embed a strings.Reader and add a Close() method that rewinds the
buffer back to the start so that another HTTP request with the same
object will be able to read the buffer.

The HTTP transport Close()s the Body after each request which is how
this hook works.