hashing: Fix bisect bugs
hashing.py that is part of the carbon project uses Python's bisect
module inconsistently.  It doesn't appear to be purposeful, however, it
uses bisect_left to locate buckets in the hash ring, but uses
bisect.insort to build the ring.  Without specifying, the insort is
bisect_right.

We add a bisectRight function and use that for building the ring.  Now
we are building identical rings even when we have multiple buckets on
the same hash code.

My example has 2700 members in the ring and the ring has 64K slots.
There are several cases of duplicate hash codes.  This seems like this
specific hash algorithm is fairly week.  Of course, improving it means
a massive re-ordering of metrics.