bucky inconsistent: Make this a first class func
This functionality will be used by future modules and needs to be in a
dedicated function.

We ignore metrics in the "carbon.agents." namespace as those are
inserted into the stream by carbon-cache after the hashing
carbon-relays.  These will never be consistent and shouldn't be.

Removed annoying log message from list.go