buckyd: up the body size limit for parsed forms
We allow POST for filtering a large list of metrics, well, its easy to
have more than 10MoB of encoded JSON data with large clusters.  So we've
upped the limit on parsed forms to 160MiB.

Most of the other API calls aren't parsed forms and therefore don't have
this limit.  Perhaps we should put the limit back but handle POST
requests with a content-type of application/json.