Special handling for TCP RSTs
If we receive an RST while waiting to read the data length for the next
incoming pickle either the client hung up on us poorly, or (in my case)
we have a TCP probe / healthcheck.  So if we receive an RST at that
point we ignore it and close the connection.  Otherwise, the error would
be logged.

Handling RST packets in Go seems odd and non-Go-like.  Not sure about
this one.