bucky-pickle-relay: timeout improvements, logging
Log errors writing to the carbon-relay-like daemon rather than blindly
re-open the connection.

Separate timeouts for "idle" connections we've accepted vs the timeout
on writing to the carbon-relay-like daemon.  A default of 30 seconds as
a timeout for pickle connections proved to be way too low in practice.
Set the default there to 5 minutes.