Fix root cause for not preserving port numbers
There are several cases where we always assume the port number is the
same, in some cases it can be different in a completely valid way.  Such
as rebalanacing from and old cluster.  This is the root cause of this,
next we just need to make sure that we preserve the port information
through the code paths.

Issue #7