Experiments with Google's Snappy compression
We don't want to spend more time compressing than we do transferring
data.  But we also have the issue that (especially with sparse whisper
files) a lot of the data is zeros or easily compressible.  So when doing
a cluster migration to another cluster whisper data is expanded to its
maximal size before shoved across the wire.  This potentially makes
transferring a new terabytes of whisper data significantly larger.