WIP: Working with compression
Code cleanup and moving more things that are common into common.go.

We now pass around the encoded version of the metric data and
decode/encode it as we need to do so.  The idea here is that this let's
us deal with normal usage where we want things decoded to write to disk,
or when we just want to pass the metric to another buckyd daemon and
don't need to do a decode/encode cycle on it.

The defaults are Identity encoding.