Unify MetricStatType and MetricData types
We also make sure to pass the X-Metric-Stat header on each request to
and from the buckyd daemon and use that metadata to help verify that the
metric data is sane.

We don't always want to decode the data in the client so we set the
encoding in the HTTP headers as normal and in the MetricData type so we
can represent the metric across the wire as a single type in memory and
only decode it when needed.