rebalance: use -metrics-per-second for reset
Before this change, if go-carbon health check is enabled, when go-carbon cache.size overflows, buckytools
reset sync speed to 1 metrics-per-second and there is no way to adjust it. This is not ideal if a go-carbon
box always has cache.size limit over the safty threshold due to file system fragmentation or aging ssd.

With this change, it allows to have an adjustable minimum sync speed.