buckyd: remove excessive usage lines from error message
on buckyd startup without paramethers it reports the error:
"You must have at least 1 node in your hash ring"
in addidtion to that it also spit a bunch of lines of usage
which are excessive and undesired as they pollute the
console or logs.
This commit is changing this behavour and only actual error
message will be returned with relatively short description.
The usage lines still can be seen by calling with --help